phase(M0.3): 探针诊断 CS0266 修复——GetCurrentThreadId 返回 IntPtr 需显式转 int（run 35086202072 实证）

+++ src/DuoFlow.App/OverlayProbe.cs
@@ -338,7 +338,7 @@ public static class OverlayProbe
             //    lum=0 (blocked UI thread starves re-composition, DD-037);
             //    Task.Delay lets the island pump messages and DWM composite
             //    while we wait.
-            int uiThread = OverlayNative.GetCurrentThreadId();
+            int uiThread = (int)OverlayNative.GetCurrentThreadId();
             Trace.Log($"probe: waiting 450ms via await Task.Delay (UI thread {uiThread} stays free)");
             await Task.Delay(450);
             Trace.Log($"probe: wait done (thread now {OverlayNative.GetCurrentThreadId()})");
